+++ packages/deploy-aws/bin/destroy-project.js
@@ -8,6 +8,15 @@ const options = require('./_project-options')
 
 process.env.AWS_REGION = options.region
 
+if (options.branch === 'master') {
+  // TODO: check confirmation
+  const confirmed = false
+  if (!confirmed) {
+    exit(1)
+    return
+  }
+}
+
 const project = new SeagullProject(options)
 project
   .destroyProject()
